Improve SMA allocation performance and lock contention
The function sma_allocate() now inserts the unused part of a split block
at the beginning of the free list, instead of reinserting it at the
position of the original block. This prevents large blocks from falling
behind in the list when many small blocks are inserted at the beginning
of the list by sma_deallocate(). Benchmarks show that this improves
allocation performance and reduces the time the SMA lock must be held.